+++ fj-doc-maven-plugin/src/main/java/org/fugerit/java/doc/project/facade/ModuleFacade.java
@@ -1,14 +1,17 @@
 package org.fugerit.java.doc.project.facade;
 
+import lombok.extern.slf4j.Slf4j;
 import org.fugerit.java.core.cfg.ConfigRuntimeException;
 
 import java.util.*;
+import java.util.stream.Collectors;
 
+@Slf4j
 public class ModuleFacade {
 
     private ModuleFacade() {}
 
-    private static final String MODULE_PREFIX = "fj-doc-";
+    public static final String MODULE_PREFIX = "fj-doc-";
 
     public static String toModuleName( String module ) {
         if ( module.contains( MODULE_PREFIX ) ) {
@@ -18,17 +21,19 @@ public class ModuleFacade {
         }
     }
 
-    private static final Set<String> MODULES = new HashSet<>( Arrays.asList(
+    private static final List<String> MODULE_LIST = Arrays.asList(
             toModuleName( "base" ),
             toModuleName( "base-json" ),
             toModuleName( "base-yaml" ),
             toModuleName( "freemarker" ),
-            toModuleName( "mod-fop" ),
             toModuleName( "mod-poi" ),
+            toModuleName( "mod-fop" ),
             toModuleName( "mod-opencsv" ),
             toModuleName( "mod-openpdf-ext" ),
             toModuleName( "mod-openrtf-ext" )
-    ) );
+    );
+
+    private static final Set<String> MODULES = new HashSet<>( MODULE_LIST );
 
     public static boolean isModuleSupported( String moduleName ) {
         return MODULES.contains( toModuleName( moduleName ) );
@@ -38,6 +43,10 @@ public class ModuleFacade {
         return Collections.unmodifiableCollection( MODULES );
     }
 
+    public static List<String> toModuleListOptimizedOrder( String extensions ) {
+        return toModuleList( extensions ).stream().sorted( Comparator.comparingInt(MODULE_LIST::indexOf) ).collect(Collectors.toList());
+    }
+
     public static List<String> toModuleList( String extensions ) {
         List<String> list = new ArrayList<>();
         for ( String currentModule : extensions.split( "," ) ) {
@@ -45,7 +54,10 @@ public class ModuleFacade {
             if ( isModuleSupported( currentModuleName ) ) {
                 list.add( currentModuleName );
             } else {
-                throw new ConfigRuntimeException( String.format( "Module not supported : %s", currentModule ) );
+                String message = String.format( "Module not supported : %s", currentModule );
+                log.warn( "{}, supported modules are : ", message );
+                ModuleFacade.getModules().forEach( log::warn );
+                throw new ConfigRuntimeException( message );
             }
         }
         return list;